# OnlineBookingAggregatorApp.Api/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/Companies/{companyId}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "operationId": "Companies_GetCompanyById",
        "parameters": [
          {
            "type": "integer",
            "name": "companyId",
            "in": "path",
            "required": true,
            "format": "int64",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CompanyModel"
            }
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Companies"
        ],
        "operationId": "Companies_UpdateCompany",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "companyId",
            "in": "path",
            "required": true,
            "format": "int64",
            "x-nullable": false
          },
          {
            "name": "model",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CompanyUpdateModel"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "422": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Companies/{companyId}/business-areas": {
      "get": {
        "tags": [
          "Companies"
        ],
        "operationId": "Companies_GetCompanyBusinessAreas",
        "parameters": [
          {
            "type": "integer",
            "name": "companyId",
            "in": "path",
            "required": true,
            "format": "int64",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BusinessAreaModel"
              }
            }
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Companies/create-company": {
      "post": {
        "tags": [
          "Companies"
        ],
        "operationId": "Companies_CreateCompany",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "name": "model",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CompanyCreateModel"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "422": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProblemDetails": {
      "type": "object",
      "additionalProperties": {},
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        },
        "extensions": {
          "type": "object",
          "additionalProperties": {}
        }
      }
    },
...
    "CompanyModel": {
      "type": "object",
      "required": [
        "id",
        "companyTypeId",
        "businessTypeId",
        "employeesSize"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "companyTypeId": {
          "$ref": "#/definitions/CompanyType"
        },
        "companyType": {
          "type": "string"
        },
        "businessTypeId": {
          "$ref": "#/definitions/BusinessType"
        },
        "employeesSize": {
          "$ref": "#/definitions/EmployeesSize"
        },
        "businessType": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "address": {
          "$ref": "#/definitions/AddressModel"
        },
        "logoPath": {
          "type": "string"
        }
      }
    },
    "CompanyType": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "JuridicalPerson",
        "PhysicalPerson"
      ],
      "enum": [
        0,
        1
      ]
    },
    "BusinessType": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Beauty",
        "Healthcare",
        "Sport",
        "Education",
        "Entertainment",
        "HouseholdServices",
        "Auto",
        "Retail",
        "Other"
      ],
      "enum": [
        1,
        2,
        3,
        4,
        5,
        6,
        7,
        8,
        9
      ]
    },
    "EmployeesSize": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "One",
        "FromTwoToFive",
        "FromSixToTen",
        "FromElevenToTwenty",
        "TwentyOnePlus"
      ],
      "enum": [
        1,
        2,
        3,
        4,
        5
      ]
    },
    "AddressModel": {
      "type": "object",
      "properties": {
        "country": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "street": {
          "type": "string"
        }
      }
    },
    "BusinessAreaModel": {
      "type": "object",
      "required": [
        "businessAreaId"
      ],
      "properties": {
        "businessAreaId": {
          "$ref": "#/definitions/BusinessArea"
        },
        "businessArea": {
          "type": "string"
        }
      }
    },
    "BusinessArea": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "BeautSalon",
        "Barbershop",
        "PrivateEmployee",
        "KidsHairdresser",
        "Spa",
        "Tattoo",
        "Cosmetology",
        "EyebrowsAndEyelashes",
        "NailService",
        "Waxing",
        "OtherBeautyServices",
        "OtherBeautyBusiness",
        "MedicalCenter",
        "Dentist",
        "VetClinic",
        "Analysis",
        "PsychotherapyAndPsychology",
        "AlternativeMedicine",
        "Hospital",
        "Polyclinic",
        "DiagnosticCenter",
        "WomenConsultation",
        "WellnessMassage",
        "OtherMedicalBusiness",
        "FitnessClub",
        "SportSchool",
        "DanceSchool",
        "TennisAndSquash",
        "SwimmingPool",
        "TrampolineCenter",
        "Ems",
        "PersonalCoach",
        "Yoga",
        "PrivateServices",
        "OtherSport",
        "MultidisciplinaryCourses",
        "KidsWorkshop",
        "LanguageSchool",
        "SchoolOfMusic",
        "TestPreparation",
        "DrivingSchool",
        "OtherEducation",
        "Sauna",
        "Bowling",
        "CyberSport",
        "Quest",
        "Restaurant",
        "MassageSalon",
        "Karting",
        "OtherEntertainment",
        "PhotoStudio",
        "Atelier",
        "DryCleaning",
        "Cleaning",
        "EquipmentRepair",
        "ZooServices",
        "EquipmentRental",
        "OtherHouseholdServices",
        "CarWash",
        "AutoBodyShop",
        "CarShowroom",
        "TuningCenter",
        "VehicleInspection",
        "DetailingCenter",
        "PrivateMechanic",
        "TyreShop",
        "CarRental",
        "AnotherAutoBusiness",
        "OpticsSalon",
        "CosmeticsStore",
        "FurnitureAndInterior",
        "FoodStore",
        "BookStore",
        "ShoppingCenter",
        "BoutiqueAndShowroom",
        "WeddingSalon",
        "OtherRetail",
        "Notary",
        "AccountingAndAuditing",
        "EventManagement",
        "LawFirm",
        "RealEstate",
        "SelfEmployedSpecialist"
      ],
      "enum": [
        1,
        2,
        3,
        4,
        5,
        6,
        7,
        8,
        9,
        10,
        11,
        12,
        13,
        14,
        15,
        16,
        17,
        18,
        19,
        20,
        21,
        22,
        23,
        24,
        25,
        26,
        27,
        28,
        29,
        30,
        31,
        32,
        33,
        34,
        35,
        36,
        37,
        38,
        39,
        40,
        41,
        42,
        43,
        44,
        45,
        46,
        47,
        48,
        49,
        50,
        51,
        52,
        53,
        54,
        55,
        56,
        57,
        58,
        59,
        60,
        61,
        62,
        63,
        64,
        65,
        66,
        67,
        68,
        69,
        70,
        71,
        72,
        73,
        74,
        75,
        76,
        77,
        78,
        79,
        80,
        81,
        82,
        83
      ]
    },
    "CompanyCreateModel": {
      "type": "object",
      "required": [
        "companyType",
        "businessType",
        "businessArea",
        "employeesSize"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "companyType": {
          "$ref": "#/definitions/CompanyType"
        },
        "businessType": {
          "$ref": "#/definitions/BusinessType"
        },
        "businessArea": {
          "$ref": "#/definitions/BusinessArea"
        },
        "employeesSize": {
          "$ref": "#/definitions/EmployeesSize"
        },
        "email": {
          "type": "string"
        },
        "address": {
          "$ref": "#/definitions/AddressDto"
        }
      }
    },
    "AddressDto": {
      "type": "object",
      "properties": {
        "country": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "street": {
          "type": "string"
        }
      }
    },
    "CompanyUpdateModel": {
      "type": "object",
      "required": [
        "companyType",
        "employeesSize"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "companyType": {
          "$ref": "#/definitions/CompanyType"
        },
        "employeesSize": {
          "$ref": "#/definitions/EmployeesSize"
        },
        "businessAreas": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BusinessArea"
          }
        },
        "email": {
          "type": "string"
        },
        "address": {
          "$ref": "#/definitions/AddressModel"
        }
      }
    },
...
  }
}