# OnlineBookingAggregatorApp.Presentation/src/app/enums/business-area.ts
export enum BusinessArea {
    BeautSalon = 1,
    Barbershop = 2,
    PrivateEmployee = 3,
    KidsHairdresser = 4,
    Spa = 5,
    Tattoo = 6,
    Cosmetology = 7,
    EyebrowsAndEyelashes = 8,
    NailService = 9,
    Waxing = 10,
    OtherBeautyServices = 11,
    OtherBeautyBusiness = 12,
    MedicalCenter = 13,
    Dentist = 14,
    VetClinic = 15,
    Analysis = 16,
    PsychotherapyAndPsychology = 17,
    AlternativeMedicine = 18,
    Hospital = 19,
    Polyclinic = 20,
    DiagnosticCenter = 21,
    WomenConsultation = 22,
    WellnessMassage = 23,
    OtherMedicalBusiness = 24,
    FitnessClub = 25,
    SportSchool = 26,
    DanceSchool = 27,
    TennisAndSquash = 28,
    SwimmingPool = 29,
    TrampolineCenter = 30,
    Ems = 31,
    PersonalCoach = 32,
    Yoga = 33,
    PrivateServices = 34,
    OtherSport = 35,
    MultidisciplinaryCourses = 36,
    KidsWorkshop = 37,
    LanguageSchool = 38,
    SchoolOfMusic = 39,
    TestPreparation = 40,
    DrivingSchool = 41,
    OtherEducation = 42,
    Sauna = 43,
    Bowling = 44,
    CyberSport = 45,
    Quest = 46,
    Restaurant = 47,
    MassageSalon = 48,
    Karting = 49,
    OtherEntertainment = 50,
    PhotoStudio = 51,
    Atelier = 52,
    DryCleaning = 53,
    Cleaning = 54,
    EquipmentRepair = 55,
    ZooServices = 56,
    EquipmentRental = 57,
    OtherHouseholdServices = 58,
    CarWash = 59,
    AutoBodyShop = 60,
    CarShowroom = 61,
    TuningCenter = 62,
    VehicleInspection = 63,
    DetailingCenter = 64,
    PrivateMechanic = 65,
    TyreShop = 66,
    CarRental = 67,
    AnotherAutoBusiness = 68,
    OpticsSalon = 69,
    CosmeticsStore = 70,
    FurnitureAndInterior = 71,
    FoodStore = 72,
    BookStore = 73,
    ShoppingCenter = 74,
    BoutiqueAndShowroom = 75,
    WeddingSalon = 76,
    OtherRetail = 77,
    Notary = 78,
    AccountingAndAuditing = 79,
    EventManagement = 80,
    LawFirm = 81,
    RealEstate = 82,
    SelfEmployedSpecialist = 83,
}